fix(new-client): missing supportsTearOut relocation
