+++ src/new-client/src/App/LiveRates/MainHeader/MainHeader.tsx
@@ -17,7 +17,7 @@ import { CurrencyOptions } from "./CurrencyOptions"
 import { currencyPairs$ } from "@/services/currencyPairs"
 import { bind } from "@react-rxjs/core"
 import { map } from "rxjs/operators"
-import { supportsTearOut } from "@/App/supportsTearOut"
+import { supportsTearOut } from "@/App/TearOutSection/supportsTearOut"
 import { TearOutComponent } from "@/App/TearOutSection/TearOutComponent"
 
 const [useCurrencies, mainHeader$] = bind(
