+++ src/new-client/src/App/TearOutSection/supportsTearOut.ts
